requirements/local.txt
asgiref==3.2.7
Django==3.0.5
django-model-utils==4.0.0
djangorestframework==3.12.4
psycopg2==2.8.6
pytz==2019.3
sqlparse==0.3.1
Unipath==1.1